Refactor: Consolidate agents into AutonomousAgent with LLM connectors
