+++ src/agents/llm_connectors/__init__.py
@@ -1,0 +1,13 @@
+"""
+LLM Connectors
+Unified interface for different LLM APIs
+"""
+
+from .base_connector import BaseLLMConnector, LLMResponse
+from .openai_connector import OpenAIConnector
+
+__all__ = [
+    "BaseLLMConnector",
+    "LLMResponse",
+    "OpenAIConnector",
+]
